Refactored MoveStrategy tests into separate class
